# 8334474: RISC-V: verify perf of ExpandBits/CompressBits (rvv)