… src/hotspot/cpu/riscv/c2_MacroAssembler_riscv.cpp
… src/hotspot/cpu/riscv/riscv_v.ad