Hi, 
Can you help to review the patch?
It removes the previous added intrinsic for Long/Integer.expand/compress, as on several real hardware, I observe obvious performance regression.

Thanks

## Performance
Tests on bananapi, for other platform, please check jbs issue for test data.

### Before
data
<google-sheets-html-origin style="caret-color: rgb(0, 0, 0); color: rgb(0, 0, 0); font-style: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: auto; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: auto; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration: none;">
Benchmark - keep intrinsic | (maxNumbits) | (size) | Mode | Cnt | Score | Error | Units
-- | -- | -- | -- | -- | -- | -- | --
o.o.b.j.lang.Integers.compress | N/A | 500 | avgt | 10 | 11710.439 | 17.936 | ns/op
o.o.b.j.lang.Integers.expand | N/A | 500 | avgt | 10 | 14878.742 | 23.472 | ns/op
o.o.b.j.lang.Longs.compress | N/A | 500 | avgt | 10 | 24555.06 | 2.632 | ns/op
o.o.b.j.lang.Longs.expand | N/A | 500 | avgt | 10 | 35827.714 | 25.022 | ns/op

</google-sheets-html-origin>


### After
data
<google-sheets-html-origin style="caret-color: rgb(0, 0, 0); color: rgb(0, 0, 0); font-style: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: auto; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: auto; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration: none;">
Benchmark - remove intrinsic | (maxNumbits) | (size) | Mode | Cnt | Score | Error | Units
-- | -- | -- | -- | -- | -- | -- | --
o.o.b.j.lang.Integers.compress | N/A | 500 | avgt | 10 | 9294.835 | 1.459 | ns/op
o.o.b.j.lang.Integers.expand | N/A | 500 | avgt | 10 | 5749.835 | 0.945 | ns/op
o.o.b.j.lang.Longs.compress | N/A | 500 | avgt | 10 | 4735.15 | 1.082 | ns/op
o.o.b.j.lang.Longs.expand | N/A | 500 | avgt | 10 | 5668.552 | 2.168 | ns/op

</google-sheets-html-origin>

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8334474](https://bugs.openjdk.org/browse/JDK-8334474): RISC-V: verify perf of ExpandBits/CompressBits (rvv) (**Sub-task** - P4)


### Reviewers
 * [Fei Yang](https://openjdk.org/census#fyang) (@RealFYang - **Reviewer**)
 * [Robbin Ehn](https://openjdk.org/census#rehn) (@robehn - **Reviewer**)
 * [Ludovic Henry](https://openjdk.org/census#luhenry) (@luhenry - Committer)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/22102/head:pull/22102` \
`$ git checkout pull/22102`

Update a local copy of the PR: \
`$ git checkout pull/22102` \
`$ git pull https://git.openjdk.org/jdk.git pull/22102/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 22102`

View PR using the GUI difftool: \
`$ git pr show -t 22102`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/22102.diff">https://git.openjdk.org/jdk/pull/22102.diff</a>

</details>
<details><summary>Using Webrev</summary>

[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/22102#issuecomment-2476152317)
</details>