@Hamlin-Li 2024-11-14
> Thanks for performing the test. Revert change looks good.

Thanks for the review. 
There was a local typo fix not commit somehow, I just added the fix, please take a another look.