@Hamlin-Li 2024-11-19
> Instead of removing, can we put it behind a flag that disabled by default? We clearly don't want to keep something that's slower for the current generation of hardware but we could expect that the next generation of hardware to go faster.

Not quite sure, how's your opinion? @RealFYang