@RealFYang 2024-11-20
> > Instead of removing, can we put it behind a flag that disabled by default? We clearly don't want to keep something that's slower for the current generation of hardware but we could expect that the next generation of hardware to go faster.
> 
> Not quite sure, as the test result shows it's too slow, and we need to introduce another vm option. How's your opinion? @RealFYang

I would suggest we revert them for maintainability. As I remembered, this code was once added when we don't have any RVV hardware to verify the performance benefit. So we just let it sit there waiting for further performance test. Now both RVV 128 
 and 256 hardwares are available and it doesn't do anything good for us on these hardwares, I don't think we would want to have it if we knew this initially. We can still adds it back if we witness performance improvement on future hardwares.